# perf: Speedup sdk.Int overflow checks