This PR speedsup all sdk.Int math ops, especially for smaller Int arithmetic.

Currently sdk.Int overflow checks compare against bigint.BitLen(), which is a complex method as it aims to be:
- constant time
- give granularity into how many bits are present in the final word. See: https://cs.opensource.google/go/go/+/refs/tags/go1.22.0:src/math/big/nat.go;l=654-674

We see this taking 1.1 seconds of the Osmosis epoch time! (Used to be 5+ seconds, until we moved more math to native bigInt math. Now its just 1.1 second in converting these final results into sdk.Int)

This PR notices that we only need to check if its greater than 256 bits, which is equivalent to checking if it takes 4 words. Note in the code, we use bits.UintSize, so it works on 32 bit machines as well.

So we change this to using int.Bits() which returns an []big.Word, with no copies, and hence will be much faster.

Notice that here in the tests we test the precise overflow bounds: https://github.com/cosmos/cosmos-sdk/blob/main/math/int_test.go#L175-L185

therefore this should be state compatible. (And I tested it with the tests under this)

Out of some abundance of caution, its possible that the most significant word could be all zeroes. To handle this edge case, we say that if the bigint uses too many words, then run the fallback (slower) BitLen logic, thus ensuring it will always be compatible.

---

## Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

I have...

* [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
* [x] confirmed `!` in the type prefix if API or client breaking change
* [x] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#pr-targeting))
* [x] provided a link to the relevant issue or specification
* [x] reviewed "Files changed" and left comments if necessary
* [x] included the necessary unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#testing)
* [x] added a changelog entry to `CHANGELOG.md`
* [x] updated the relevant documentation or specification, including comments for [documenting Go code](https://blog.golang.org/godoc)
* [x] confirmed all CI checks have passed

## Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

I have...

* [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
* [ ] confirmed all author checklist items have been addressed
* [ ] reviewed state machine logic, API design and naming, documentation is accurate, tests and test coverage


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **Refactor**
	- Enhanced performance and reliability by improving the handling of maximum word length for integer types.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->