@ValarDragon 2024-02-08
Hrmm, maybe this isn't a great idea though. Trying to think of anything that could potentially lead to an empty most-significant-word, which would then make this potentially not work. I guess I could make a method that does some fall back logic if len(i.Bits()) > 4. (Since that should almost never be the case, and its fine to do more logic in that situation)