@ValarDragon 2024-02-08
I added a benchmark to show the speed difference on the existing table of small values (<= 78 bits). I expect this difference becomes more pronounced on the real Osmosis data set.

Before:
```
BenchmarkIntOverflowCheckTime-12    	 4743949	       228.5 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	cosmossdk.io/math	1.437s
```

After:
```
BenchmarkIntOverflowCheckTime-12    	 6575220	       175.3 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	cosmossdk.io/math	1.557s
```