@ValarDragon 2024-02-09
TYSM! I also learned this morning, that its impossible for the top word of a Nat to have 0's, so we could omit the fallback check -- but it doesn't really matter since its rarely called.